test 6 and 7, solution: line pointer array
