+++ input.txt
@@ -1,0 +1,9 @@
+Hola
+a todos
+mi nombre
+es
+sebastian
+a
+a
+a
+a
