+++ tests-out/6.other
@@ -1,0 +1,5 @@
+test
+a
+is
+this
+hello
